TM-1398: powershell: more whatif fixes (#1486)
* test

* -

* -

* -

* -

* -

* -

* -

* -

* fix

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

* -

+++ powershell/Scripts/Microsoft/Add-EdgePopupsAllowedForUrls.ps1
@@ -21,21 +21,25 @@ if (!(Test-Path $RegPath)) {
   New-Item -Path $RegPath -Force | Out-Null
 }
 
-$ItemName = 1
-foreach ($PopupsAllowedForUrl in $PopupsAllowedForUrls) {
-  $ItemProperty = Get-ItemProperty -Path $RegPath -Name $ItemName -ErrorAction SilentlyContinue
-  if ($null -eq $ItemProperty -or $ItemProperty.$ItemName -ne $PopupsAllowedForUrl) {
-    Write-Output "Setting $RegPath\$ItemName = $PopupsAllowedForUrl"
-    New-ItemProperty -Path $RegPath -Name $ItemName -Value $PopupsAllowedForUrl -PropertyType String -Force | Out-Null
+# Add check here in case $WhatIfPreference has been used
+if (Test-Path $RegPath) {
+  $ItemName = 1
+  foreach ($PopupsAllowedForUrl in $PopupsAllowedForUrls) {
+    $ItemProperty = Get-ItemProperty -Path $RegPath -Name $ItemName -ErrorAction SilentlyContinue
+    if ($null -eq $ItemProperty -or $ItemProperty.$ItemName -ne $PopupsAllowedForUrl) {
+      Write-Output "Setting $RegPath\$ItemName = $PopupsAllowedForUrl"
+      New-ItemProperty -Path $RegPath -Name $ItemName -Value $PopupsAllowedForUrl -PropertyType String -Force | Out-Null
+    }
+    $ItemName = $ItemName + 1
   }
-  $ItemName = $ItemName + 1
-}
 
-$ItemProperty = Get-ItemProperty -Path $RegPath -Name $ItemName -ErrorAction SilentlyContinue
-while ($null -ne $ItemProperty) {
-  $PopupsAllowedForUrl = $ItemProperty.$ItemName
-  Write-Output "Removing $RegPath\$ItemName = $PopupsAllowedForUrl"
-  Remove-ItemProperty -Path $RegPath -Name $ItemName | Out-Null
-  $ItemName = $ItemName + 1
+
   $ItemProperty = Get-ItemProperty -Path $RegPath -Name $ItemName -ErrorAction SilentlyContinue
+  while ($null -ne $ItemProperty) {
+    $PopupsAllowedForUrl = $ItemProperty.$ItemName
+    Write-Output "Removing $RegPath\$ItemName = $PopupsAllowedForUrl"
+    Remove-ItemProperty -Path $RegPath -Name $ItemName | Out-Null
+    $ItemName = $ItemName + 1
+    $ItemProperty = Get-ItemProperty -Path $RegPath -Name $ItemName -ErrorAction SilentlyContinue
+  }
 }
